+++ lib/logarte.dart
@@ -5,5 +5,5 @@ export 'src/models/logarte_entry.dart';
 export 'src/channels/logarte_navigator_observer.dart';
 export 'src/channels/logarte_dio_interceptor.dart';
 export 'src/console/logarte_magical_tap.dart';
-export 'src/console/logarte_magical_tap.dart';
 export 'src/console/logarte_dashboard_screen.dart';
+export 'src/logger/level.dart';
