+++ lib/src/logger/printers/prefix_printer.dart
@@ -1,5 +1,6 @@
 import 'package:logarte/src/logger/log_printer.dart';
 import 'package:logarte/src/logger/logger.dart';
+import 'package:logarte/src/logger/level.dart';
 
 /// A decorator for a [LogPrinter] that allows for the prepending of every
 /// line in the log output with a string for the level of that log. For
