+++ lib/src/logger/printers/simple_printer.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-
+import 'package:logarte/src/logger/level.dart';
 import 'package:logarte/src/logger/ansi_color.dart';
 import 'package:logarte/src/logger/log_printer.dart';
 import 'package:logarte/src/logger/logger.dart';
